saveImageToPrivateStorage moved from PlaylistFragment to PlaylistViewModel
