+++ app/src/main/java/com/example/yp_playlist_maker/di/AppModule.kt
@@ -48,7 +48,9 @@ val appModule = module {
 
     viewModel<PlaylistViewModel> {
         PlaylistViewModel(
-            playlistsInteractor = get()
+            playlistsInteractor = get(),
+            playlistsDirectory = get(),
+            contentResolver = get()
         )
     }
